requirements.txt
fastapi
uvicorn[standard]
streamlit
requests

langchain
langchain-community

sentence-transformers
chromadb

pandas
python-dotenv

pyjwt
passlib
python-jose[cryptography]

httpx
pydantic

python-multipart

openai
groq
